handling session timeout (#185)
* handling session timeout

* csproj

* changing redirect URL

* changed redirection

* redirection changes

* handling session expiry

* increase in session timeout

* error message on session expiry

* error handling

+++ src/VstsDemoBuilder/Controllers/EnvironmentController.cs
@@ -34,6 +34,7 @@ namespace VstsDemoBuilder.Controllers
 
         [HttpGet]
         [AllowAnonymous]
+        [SessonTimeout]
         public ContentResult GetCurrentProgress(string id)
         {
             this.ControllerContext.HttpContext.Response.AddHeader("cache-control", "no-cache");
@@ -48,6 +49,7 @@ namespace VstsDemoBuilder.Controllers
         /// <returns></returns>
         [HttpGet]
         [AllowAnonymous]
+        [SessonTimeout]
         public string GetStatusMessage(string id)
         {
             lock (ProjectService.objLock)
@@ -78,6 +80,7 @@ namespace VstsDemoBuilder.Controllers
         /// <returns></returns>
         [HttpGet]
         [AllowAnonymous]
+        [SessonTimeout]
         public ContentResult GetTemplate(string TemplateName)
         {
             string templatesPath = Server.MapPath("~") + @"\Templates\";
@@ -97,6 +100,7 @@ namespace VstsDemoBuilder.Controllers
         /// <param name="model"></param>
         /// <returns></returns>
         [AllowAnonymous]
+        [SessonTimeout]
         public JsonResult GetGroups()
         {
             string groupDetails = "";
@@ -264,6 +268,7 @@ namespace VstsDemoBuilder.Controllers
             }
         }
         [AllowAnonymous]
+        [SessonTimeout]
         public ActionResult PrivateTemplate()
         {
             if (Session["visited"] != null)
@@ -328,6 +333,7 @@ namespace VstsDemoBuilder.Controllers
 
         [AllowAnonymous]
         [HttpPost]
+        [SessonTimeout]
         public ActionResult UploadFiles()
         {
             string[] strResult = new string[2];
@@ -398,6 +404,7 @@ namespace VstsDemoBuilder.Controllers
 
         [AllowAnonymous]
         [HttpPost]
+        [SessonTimeout]
         public ActionResult UnzipFile(string fineName)
         {
             PrivateTemplate privateTemplate = new PrivateTemplate();
@@ -461,6 +468,7 @@ namespace VstsDemoBuilder.Controllers
         /// <returns></returns>
         [HttpPost]
         [AllowAnonymous]
+        [SessonTimeout]
         public JsonResult GetMembers(string accountName, string accessToken)
         {
             Project mod = new Project();
@@ -497,24 +505,32 @@ namespace VstsDemoBuilder.Controllers
         /// <returns></returns>
         [HttpPost]
         [AllowAnonymous]
+        [SessonTimeout]
         public bool StartEnvironmentSetupProcess(Project model)
         {
             try
             {
-                Session["PAT"] = model.accessToken;
-                Session["AccountName"] = model.accountName;
-                if (Session["GitHubToken"] != null && Session["GitHubToken"].ToString() != "" && model.GitHubFork)
+                if (Session["visited"] != null)
                 {
-                    model.GitHubToken = Session["GitHubToken"].ToString();
+                    Session["PAT"] = model.accessToken;
+                    Session["AccountName"] = model.accountName;
+                    if (Session["GitHubToken"] != null && Session["GitHubToken"].ToString() != "" && model.GitHubFork)
+                    {
+                        model.GitHubToken = Session["GitHubToken"].ToString();
+                    }
+                    projectService.AddMessage(model.id, string.Empty);
+                    projectService.AddMessage(model.id.ErrorId(), string.Empty);
+                    if (!string.IsNullOrEmpty(model.PrivateTemplatePath))
+                    {
+                        model.IsPrivatePath = true;
+                    }
+                    ProcessEnvironment processTask = new ProcessEnvironment(projectService.CreateProjectEnvironment);
+                    processTask.BeginInvoke(model, new AsyncCallback(EndEnvironmentSetupProcess), processTask);
                 }
-                projectService.AddMessage(model.id, string.Empty);
-                projectService.AddMessage(model.id.ErrorId(), string.Empty);
-                if (!string.IsNullOrEmpty(model.PrivateTemplatePath))
+                else
                 {
-                    model.IsPrivatePath = true;
+                    return false;
                 }
-                ProcessEnvironment processTask = new ProcessEnvironment(projectService.CreateProjectEnvironment);
-                processTask.BeginInvoke(model, new AsyncCallback(EndEnvironmentSetupProcess), processTask);
             }
             catch (Exception ex)
             {
@@ -599,6 +615,7 @@ namespace VstsDemoBuilder.Controllers
         /// <returns></returns>
         [HttpGet]
         [AllowAnonymous]
+        [SessonTimeout]
         public JsonResult CheckForInstalledExtensions(string selectedTemplate, string token, string account, string PrivatePath = "")
         {
             try
@@ -759,6 +776,7 @@ namespace VstsDemoBuilder.Controllers
         }
 
         [AllowAnonymous]
+        [SessonTimeout]
         public string CheckSession()
         {
             if (Session["GitHubToken"] != null && Session["GitHubToken"].ToString() != "")
@@ -773,49 +791,59 @@ namespace VstsDemoBuilder.Controllers
 
         [HttpGet]
         [AllowAnonymous]
+        [SessonTimeout]
         public JsonResult UploadPrivateTemplateFromURL(string TemplateURL, string token, string userId, string password, string OldPrivateTemplate = "")
         {
-            if (!string.IsNullOrEmpty(OldPrivateTemplate))
+            if (Session["visited"] != null)
             {
-                templateService.deletePrivateTemplate(OldPrivateTemplate);
-            }
-            PrivateTemplate privateTemplate = new PrivateTemplate();
-            string templatePath = string.Empty;
-            try
-            {
-                string templateName = "";
-                string fileName = Path.GetFileName(TemplateURL);
-                string extension = Path.GetExtension(TemplateURL);
-                templateName = fileName.ToLower().Replace(".zip", "").Trim() + "-" + Guid.NewGuid().ToString().Substring(0, 6) + extension.ToLower();
-                privateTemplate.privateTemplateName = templateName.ToLower().Replace(".zip", "").Trim();
-                privateTemplate.privateTemplatePath = templateService.GetTemplateFromPath(TemplateURL, templateName, token, userId, password);
-
-                if (privateTemplate.privateTemplatePath != "")
+                if (!string.IsNullOrEmpty(OldPrivateTemplate))
                 {
-                    privateTemplate.responseMessage = templateService.checkSelectedTemplateIsPrivate(privateTemplate.privateTemplatePath);
-                    if (privateTemplate.responseMessage != "SUCCESS")
+                    templateService.deletePrivateTemplate(OldPrivateTemplate);
+                }
+                PrivateTemplate privateTemplate = new PrivateTemplate();
+                string templatePath = string.Empty;
+                try
+                {
+                    string templateName = "";
+                    string fileName = Path.GetFileName(TemplateURL);
+                    string extension = Path.GetExtension(TemplateURL);
+                    templateName = fileName.ToLower().Replace(".zip", "").Trim() + "-" + Guid.NewGuid().ToString().Substring(0, 6) + extension.ToLower();
+                    privateTemplate.privateTemplateName = templateName.ToLower().Replace(".zip", "").Trim();
+                    privateTemplate.privateTemplatePath = templateService.GetTemplateFromPath(TemplateURL, templateName, token, userId, password);
+
+                    if (privateTemplate.privateTemplatePath != "")
                     {
-                        var templatepath = HostingEnvironment.MapPath("~") + @"\PrivateTemplates\" + templateName.ToLower().Replace(".zip", "").Trim();
-                        if (Directory.Exists(templatepath))
-                            Directory.Delete(templatepath, true);
+                        privateTemplate.responseMessage = templateService.checkSelectedTemplateIsPrivate(privateTemplate.privateTemplatePath);
+                        if (privateTemplate.responseMessage != "SUCCESS")
+                        {
+                            var templatepath = HostingEnvironment.MapPath("~") + @"\PrivateTemplates\" + templateName.ToLower().Replace(".zip", "").Trim();
+                            if (Directory.Exists(templatepath))
+                                Directory.Delete(templatepath, true);
+                        }
                     }
-                }
-                else
-                {
-                    privateTemplate.responseMessage = "Unable to download file, please check the provided URL";
-                }
+                    else
+                    {
+                        privateTemplate.responseMessage = "Unable to download file, please check the provided URL";
+                    }
 
+                }
+                catch (Exception ex)
+                {
+                    ProjectService.logger.Info(DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss") + "\t" + "\t" + ex.Message + "\t" + "\n" + ex.StackTrace + "\n");
+                    return Json(new { message = "Error", status = "false" }, JsonRequestBehavior.AllowGet);
+                }
+                return Json(privateTemplate, JsonRequestBehavior.AllowGet);
             }
-            catch (Exception ex)
+            else
             {
-                ProjectService.logger.Info(DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss") + "\t" + "\t" + ex.Message + "\t" + "\n" + ex.StackTrace + "\n");
-                return Json(new { message = "Error", status = "false" }, JsonRequestBehavior.AllowGet);
+                return Json("Session Expired", JsonRequestBehavior.AllowGet);
             }
-            return Json(privateTemplate, JsonRequestBehavior.AllowGet);
+
         }
 
         [HttpPost]
         [AllowAnonymous]
+        [SessonTimeout]
         public void DeletePrivateTemplate(string TemplateName)
         {
             templateService.deletePrivateTemplate(TemplateName);
